+++ q01_10/007.py
@@ -4,7 +4,7 @@ import lib0110
 
 if __name__ == '__main__':
     img = cv2.imread("Gasyori100knock/Question_01_10/imori.jpg")
-    ans = lib0110.mean_pooling(img,8).astype(np.uint8)
+    ans = lib0110.mean_pooling(img,8)
 
     cv2.imshow("imori", ans)
     cv2.waitKey(0)
