+++ q11_20/011.py
@@ -4,7 +4,7 @@ import lib1120
 
 if __name__ == '__main__':
     img = cv2.imread("Gasyori100knock/Question_11_20/imori.jpg")
-    ans = lib1120.Smoothing_filter(img).astype(np.uint8)
+    ans = lib1120.Smoothing_filter(img)
 
     cv2.imshow("imori", ans)
     cv2.waitKey(0)
